rt: flatten module file structure
