--- crates/bevy-realtime/src/internal/mod.rs
@@ -1,4 +1,0 @@
-pub mod channel;
-pub mod client;
-pub mod message;
-pub mod presence;
