+++ crates/bevy-realtime/src/postgres_changes.rs
@@ -1,71 +1,80 @@
-use std::marker::PhantomData;
+pub mod bevy {
+    use std::marker::PhantomData;
 
-use bevy::prelude::*;
-use crossbeam::channel::unbounded;
+    use bevy::prelude::*;
+    use crossbeam::channel::unbounded;
 
-use crate::{
-    forwarder_recv,
-    internal::message::{
-        payload::{PostgresChangesEvent, PostgresChangesPayload},
-        postgres_change_filter::PostgresChangeFilter,
-    },
-    BevyChannelBuilder, ChannelForwarder,
-};
+    use crate::{
+        forwarder_recv,
+        message::{
+            payload::{PostgresChangesEvent, PostgresChangesPayload},
+            postgres_change_filter::PostgresChangeFilter,
+        },
+        BevyChannelBuilder, ChannelForwarder,
+    };
 
-pub trait PostgresPayloadEvent {
-    fn new(payload: PostgresChangesPayload) -> Self;
-}
-
-pub trait AppExtend {
-    fn add_postgres_event<E: Event + PostgresPayloadEvent, F: Component>(&mut self) -> &mut Self;
-}
-
-impl AppExtend for App {
-    fn add_postgres_event<E: Event + PostgresPayloadEvent, F: Component>(&mut self) -> &mut Self {
-        self.add_event::<E>().add_systems(
-            Update,
-            (postgres_forward::<E, F>, forwarder_recv::<E>).chain(),
-        )
+    pub trait PostgresPayloadEvent {
+        fn new(payload: PostgresChangesPayload) -> Self;
     }
-}
 
-#[derive(Component)]
-pub struct PostgresForwarder<E: Event + PostgresPayloadEvent> {
-    pub event: PostgresChangesEvent,
-    pub filter: PostgresChangeFilter,
-    spooky: PhantomData<E>,
-}
+    pub trait AppExtend {
+        fn add_postgres_event<E: Event + PostgresPayloadEvent, F: Component>(
+            &mut self,
+        ) -> &mut Self;
+    }
 
-impl<E: Event + PostgresPayloadEvent> PostgresForwarder<E> {
-    pub fn new(event: PostgresChangesEvent, filter: PostgresChangeFilter) -> PostgresForwarder<E> {
-        Self {
-            event,
-            filter,
-            spooky: PhantomData::<E>,
+    impl AppExtend for App {
+        fn add_postgres_event<E: Event + PostgresPayloadEvent, F: Component>(
+            &mut self,
+        ) -> &mut Self {
+            self.add_event::<E>().add_systems(
+                Update,
+                (postgres_forward::<E, F>, forwarder_recv::<E>).chain(),
+            )
         }
     }
-}
 
-// "consumes" PostgresForwarders, creates ChannelForwarders
-pub fn postgres_forward<E: Event + PostgresPayloadEvent, T: Component>(
-    mut commands: Commands,
-    mut q: Query<
-        (Entity, &mut BevyChannelBuilder, &PostgresForwarder<E>),
-        (Added<PostgresForwarder<E>>, With<T>),
-    >,
-) {
-    for (e, mut cb, event) in q.iter_mut() {
-        let (tx, rx) = unbounded();
+    #[derive(Component)]
+    pub struct PostgresForwarder<E: Event + PostgresPayloadEvent> {
+        pub event: PostgresChangesEvent,
+        pub filter: PostgresChangeFilter,
+        spooky: PhantomData<E>,
+    }
 
-        cb.0.on_postgres_change(event.event.clone(), event.filter.clone(), move |payload| {
-            let ev = E::new(payload.clone());
+    impl<E: Event + PostgresPayloadEvent> PostgresForwarder<E> {
+        pub fn new(
+            event: PostgresChangesEvent,
+            filter: PostgresChangeFilter,
+        ) -> PostgresForwarder<E> {
+            Self {
+                event,
+                filter,
+                spooky: PhantomData::<E>,
+            }
+        }
+    }
 
-            tx.try_send(ev).unwrap();
-        });
+    // "consumes" PostgresForwarders, creates ChannelForwarders
+    pub fn postgres_forward<E: Event + PostgresPayloadEvent, T: Component>(
+        mut commands: Commands,
+        mut q: Query<
+            (Entity, &mut BevyChannelBuilder, &PostgresForwarder<E>),
+            (Added<PostgresForwarder<E>>, With<T>),
+        >,
+    ) {
+        for (e, mut cb, event) in q.iter_mut() {
+            let (tx, rx) = unbounded();
 
-        commands
-            .entity(e)
-            .insert(ChannelForwarder::<E> { rx })
-            .remove::<PostgresForwarder<E>>();
+            cb.0.on_postgres_change(event.event.clone(), event.filter.clone(), move |payload| {
+                let ev = E::new(payload.clone());
+
+                tx.try_send(ev).unwrap();
+            });
+
+            commands
+                .entity(e)
+                .insert(ChannelForwarder::<E> { rx })
+                .remove::<PostgresForwarder<E>>();
+        }
     }
 }
